added quality measures

+++ src/main/java/org/example/model/UniverseOfDiscourse.java
@@ -5,18 +5,18 @@ import lombok.Getter;
 @Getter
 public class UniverseOfDiscourse {
 
-    private final Integer minimum;
-    private final Integer maximum;
-    private final Integer size;
+    private final Double minimum;
+    private final Double maximum;
+    private final Double size;
 
-    public UniverseOfDiscourse(Integer minimum, Integer maximum) {
+    public UniverseOfDiscourse(Double minimum, Double maximum) {
         if (maximum - minimum <= 0) throw new IllegalStateException("Invalid universe of discourse range!");
         this.size = maximum - minimum;
         this.minimum = minimum;
         this.maximum = maximum;
     }
 
-    public boolean valueInUniverseOfDiscourse(Integer value) {
+    public boolean valueInUniverseOfDiscourse(Double value) {
         return value <= minimum && value >= maximum;
     }
 }
